docs: add schedule api description & fix rest api route
